+++ src/dirmodel.h
@@ -3,7 +3,7 @@
 
 #include "listmodel.h"
 
-const char *dirmodel_getfilename(struct listmodel *model, unsigned int index);
+const char *dirmodel_getfilename(struct listmodel *model, size_t index);
 void dirmodel_change_directory(struct listmodel *model, const char *path);
 void dirmodel_init(struct listmodel *model, const char *path);
 void dirmodel_free(struct listmodel *model);
